+++ bin/opup/Cargo.toml
@@ -11,6 +11,7 @@ license.workspace = true
 op-config = { path = "../../crates/config" }
 op-primitives = { path = "../../crates/primitives" }
 op-contracts = { path = "../../crates/contracts" }
+op-stages = { path = "../../crates/stages" }
 
 serde.workspace = true
 serde_json.workspace = true
@@ -20,7 +21,6 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 
 bollard = "0.15.0"
-project-root = "0.2.2"
 tokio = { version = "1.11.0", features = ["full"] }
 clap = { version = "4.4.3", features = ["derive"] }
