+++ bin/opup/src/cli.rs
@@ -1,5 +1,6 @@
 use clap::{ArgAction, Parser};
 use eyre::Result;
+use std::path::PathBuf;
 
 /// Command line arguments
 #[derive(Parser, Debug)]
@@ -8,16 +9,22 @@ pub struct Args {
     /// Verbosity level (0-4)
     #[arg(long, short, action = ArgAction::Count, default_value = "2")]
     v: u8,
+    /// An optional path to a stack config file.
+    #[arg(long, short)]
+    config: Option<PathBuf>,
 }
 
 pub fn run() -> Result<()> {
-    let Args { v } = Args::parse();
+    let Args { v, config } = Args::parse();
 
     crate::telemetry::init_tracing_subscriber(v)?;
 
-    tracing::info!(target: "opup", "bootstrapping op stack");
+    crate::banners::banner()?;
 
-    crate::stack::run()?;
+    // todo: switch on subcommands
+    // default should be to run the devnet stack
+    // should also allow nuking the devnet
+    // and stopping an already running devnet
 
-    Ok(())
+    crate::stack::Stack::new(config).run()
 }
