--- bin/opup/src/constants.rs
@@ -1,27 +1,0 @@
-/// Optimism monorepo git url.
-pub const OP_MONOREPO_URL: &str = "git@github.com:ethereum-optimism/optimism.git";
-
-/// L1 node url.
-pub const L1_URL: &str = "http://localhost:8545";
-
-/// L1 node port.
-pub const L1_PORT: u16 = 8545;
-
-/// L2 node url.
-pub const L2_URL: &str = "http://localhost:9545";
-
-/// L2 node port.
-pub const L2_PORT: u16 = 9545;
-
-/// Rollup node url.
-pub const ROLLUP_URL: &str = "http://localhost:7545";
-
-/// Rollup node port.
-pub const ROLLUP_PORT: u16 = 7545;
-
-/// Testing deployer private key.
-pub const DEPLOYER_PRIVATE_KEY: &str =
-    "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
-
-/// Zero address string constant with a `0x` prefix.
-pub const ADDRESS_ZERO: &str = "0x0000000000000000000000000000000000000000";
