--- bin/opup/src/etc/mod.rs
@@ -1,14 +1,0 @@
-/// Runner contains asynchronous helpers for running commands.
-pub mod runner;
-
-/// Contains helper functions for shell commands.
-pub mod commands;
-
-/// Contains helper functions for git commands.
-pub mod git;
-
-/// Contains helper functions for networking.
-pub mod net;
-
-/// Contains JSON helpers.
-pub mod json;
