+++ bin/opup/src/lib.rs
@@ -4,14 +4,11 @@ pub mod cli;
 /// Module containing stack packages.
 pub mod stack;
 
-/// General purpose helpers.
-pub mod etc;
-
-/// System-level constants.
-pub mod constants;
-
 /// Command banners.
 pub mod banners;
 
 /// Telemetry contains helpers for initializing telemetry.
 pub mod telemetry;
+
+/// Runner contains asynchronous helpers for running commands.
+pub mod runner;
