+++ crates/stages/Cargo.toml
@@ -11,8 +11,13 @@ homepage.workspace = true
 repository.workspace = true
 
 [dependencies]
+op-contracts = { path = "../contracts" }
+op-config = { path = "../config" }
 op-primitives = { path = "../primitives" }
 
 eyre.workspace = true
 tracing.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 
+project-root = "0.2.2"
