+++ crates/stages/src/commands.rs
@@ -3,7 +3,7 @@ use std::path::Path;
 use std::process::{Command, Output};
 
 /// Check command checks the output of a command and returns an error if it failed.
-pub fn check_command(out: Output, err: &str) -> Result<()> {
+pub(crate) fn check_command(out: Output, err: &str) -> Result<()> {
     if !out.status.success() {
         eyre::bail!(
             "Failed to run command: {}: {}",
@@ -14,7 +14,9 @@ pub fn check_command(out: Output, err: &str) -> Result<()> {
     Ok(())
 }
 
-pub fn make_executable(path: &Path) -> Result<()> {
+/// Makes a file executable.
+#[allow(dead_code)]
+pub(crate) fn make_executable(path: &Path) -> Result<()> {
     let path_str = path.to_str().expect("Failed to convert path to string");
     let out = Command::new("chmod").args(["+x", path_str]).output()?;
     check_command(out, &format!("Failed to make {} executable", path_str))?;
