+++ crates/stages/src/git.rs
@@ -1,9 +1,9 @@
-use crate::etc::commands::check_command;
+use crate::commands::check_command;
 use eyre::Result;
 use std::{path::Path, process::Command};
 
 /// Clones a given git repository into the given directory.
-pub fn git_clone(pwd: &Path, repo: &str) -> Result<()> {
+pub(crate) fn git_clone(pwd: &Path, repo: &str) -> Result<()> {
     let out = Command::new("git")
         .arg("clone")
         .arg("--recursive")
