+++ crates/stages/src/net.rs
@@ -4,7 +4,7 @@ use std::thread;
 use std::time::Duration;
 
 /// Wait for a port to come up.
-pub fn wait_up(port: u16, retries: u32, wait_secs: u64) -> Result<()> {
+pub(crate) fn wait_up(port: u16, retries: u32, wait_secs: u64) -> Result<()> {
     for _ in 0..retries {
         tracing::debug!(target: "opup", "Trying 127.0.0.1:{}", port);
         if let Ok(stream) = TcpStream::connect(SocketAddr::from(([127, 0, 0, 1], port))) {
